Update lock tab counts every 1s with single timer
Track the timer in a context shared by all <OpenTabRow> components.
